add path following working!
the vehicle will follow the path by checking if it's inside any polygon and if not by taking the normal point + an offset on the nearest segment with a starting x greater than his. The last constraint is critical to have a vehicle simulating following a path, else the vehicle would go back and forth.

Also fixed a small issue in the vehicle logic.